feat: additional tests added

+++ ldi-nifi/ldi-nifi-processors/ldes-client-processor/src/main/java/be/vlaanderen/informatievlaanderen/ldes/ldi/processors/LdesClientProcessor.java
@@ -182,7 +182,7 @@ public class LdesClientProcessor extends AbstractProcessor {
 
 	@OnRemoved
 	public void onRemoved() {
-		if (!keepState) {
+		if (!keepState && memberSupplier != null) {
 			memberSupplier.destroyState();
 		}
 	}
